feat: keymaps for color selection (#251)
This commit adds the following keymaps:

- P: toggle the color palette UI
- 0: select palette color 0
- 1: select palette color 1
- 2: select palette color 2
- 3: select palette color 3
- 4: select palette color 4
- 5: select palette color 5
- 6: select palette color 6
- 7: select palette color 7
- 8: select palette color 8
- 9: select palette color 9

If there are fewer than 10 colors in the palette, keymaps for colors
beyond what's in the palette will result in selecting the last color.
For example, if there are 3 colors in the palette and the user presses
4, then color 3 is selected.

For palettes with more than 10 colors, manual selection is still
required for colors after 10.